add data, after deleting old big objects
---
On the future data should be
on a server, on a storage thing.

However, this project is small,
and the data is big in relation
so source files, it is not big
on terms of fast download.

Given a copy of data.tar.gz,
we could remove all old data.tar.gz objects like this:

``` sh
    git clone https://github.com/elbecerrasoto/PreToxin FreshPreToxin
    cd FreshPreToxin
    git filter-repo --path data.tar.gz --invert-paths
    git remote add origin https://github.com/elbecerrasoto/PreToxin.git
    git push origin --force --all
```